backend/requirements.txt
Django>=5.0,<6
django-ninja>=1.3,<2
django-cors-headers>=4.3,<5
python-dotenv>=1.0,<2
